refactor: split Utils into multiple files
