+++ Editor/Utils/AnimatorLayerMap.cs
@@ -1,0 +1,40 @@
+using System;
+using VRC.SDK3.Avatars.Components;
+
+namespace Anatawa12.AvatarOptimizer
+{
+    class AnimatorLayerMap<T>
+    {
+        private T[] _values = new T[(int)(VRCAvatarDescriptor.AnimLayerType.IKPose + 1)];
+
+        public static bool IsValid(VRCAvatarDescriptor.AnimLayerType type)
+        {
+            switch (type)
+            {
+                case VRCAvatarDescriptor.AnimLayerType.Base:
+                case VRCAvatarDescriptor.AnimLayerType.Additive:
+                case VRCAvatarDescriptor.AnimLayerType.Gesture:
+                case VRCAvatarDescriptor.AnimLayerType.Action:
+                case VRCAvatarDescriptor.AnimLayerType.FX:
+                case VRCAvatarDescriptor.AnimLayerType.Sitting:
+                case VRCAvatarDescriptor.AnimLayerType.TPose:
+                case VRCAvatarDescriptor.AnimLayerType.IKPose:
+                    return true;
+                case VRCAvatarDescriptor.AnimLayerType.Deprecated0:
+                default:
+                    return false;
+            }
+        }
+
+        public ref T this[VRCAvatarDescriptor.AnimLayerType type]
+        {
+            get
+            {
+                if (!IsValid(type))
+                    throw new ArgumentOutOfRangeException(nameof(type), type, null);
+
+                return ref _values[(int)type];
+            }
+        }
+    }
+}
